Moved socal auth routes to their own controller
Also cleaned some phpdocs and extracted register actions to their own
service.
